requirements.txt
fastapi==0.115.6
numpy==1.26.4
Pillow==11.0.0
tflite_runtime==2.14.0
uvicorn==0.32.1
python-multipart==0.0.19